addContactMiddleware
