+++ routes/api/contacts.js
@@ -3,6 +3,10 @@ const router = express.Router()
 
 const { authMiddleware } = require('../../middleware/authMiddleware');
 const {
+  addContactMiddleware
+} = require('../../middleware/contactMiddleware');
+
+const {
   getContactsController,
   getContactByIdController,
   addContactController,
@@ -14,7 +18,7 @@ const {
 router.use(authMiddleware);
 
 router.get('/', getContactsController);
-router.post('/', addContactController);
+router.post('/', addContactMiddleware, addContactController);
 router.get('/:contactId', getContactByIdController);
 router.put('/:contactId', updateContactController);
 router.delete('/:contactId', removeContactController);
